refactor(voices): cleanup unused imports and improve layout structure for campaign and proposals sections
