+++ catalyst_voices/apps/voices/lib/pages/discovery/sections/campaign_categories.dart
@@ -1,6 +1,4 @@
-import 'package:catalyst_voices/common/ext/build_context_ext.dart';
 import 'package:catalyst_voices/widgets/cards/campaign_category_card.dart';
-import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
 import 'package:catalyst_voices_view_models/catalyst_voices_view_models.dart';
 import 'package:flutter/material.dart';
 import 'package:skeletonizer/skeletonizer.dart';
@@ -23,11 +21,6 @@ class CampaignCategories extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          Text(
-            context.l10n.campaignCategories,
-            style: context.textTheme.titleLarge,
-          ),
-          const SizedBox(height: 24),
           Wrap(
             spacing: 12,
             runSpacing: 12,
