+++ catalyst_voices/apps/voices/lib/pages/discovery/state_selectors/campaign_categories_state_selector.dart
@@ -1,3 +1,4 @@
+import 'package:catalyst_voices/common/ext/build_context_ext.dart';
 import 'package:catalyst_voices/pages/discovery/sections/campaign_categories.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
@@ -14,11 +15,25 @@ class CampaignCategoriesStateSelector extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Stack(
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        _CampaignCategoriesLoading(),
-        _CampaignCategoriesData(),
-        _CampaignCategoriesError(),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 120),
+          child: Text(
+            context.l10n.campaignCategories,
+            style: context.textTheme.titleLarge,
+          ),
+        ),
+        const SizedBox(height: 24),
+        const Stack(
+          children: [
+            _CampaignCategoriesLoading(),
+            _CampaignCategoriesData(),
+            _CampaignCategoriesError(),
+          ],
+        ),
       ],
     );
   }
@@ -40,7 +55,7 @@ class _CampaignCategoriesLoading extends StatelessWidget {
         );
         return Offstage(
           offstage: !state,
-          child: CampaignCategories(dummyCategories, isLoading: true),
+          child: CampaignCategories(dummyCategories, isLoading: state),
         );
       },
     );
